+++ demos/48-contour-plot/index.js
@@ -21,6 +21,7 @@
 
   window.addEventListener("load", function () {
     var isDarkmode = detectDarkMode(GUP);
+    var params = new Params(GUP);
 
     // All config params are optional.
     var pb = new PlotBoilerplate(
@@ -58,6 +59,9 @@
     // +---------------------------------------------------------------------------------
     // | A global config that's attached to the dat.gui control interface.
     // +-------------------------------
+    var CLOSE_GAP_TYPE_NONE = 0;
+    var CLOSE_GAP_TYPE_ABOVE = 1;
+    var CLOSE_GAP_TYPE_BELOW = 2;
     var config = PlotBoilerplate.utils.safeMergeByKeys(
       {
         showNormals: false,
@@ -69,11 +73,13 @@
         },
 
         sliceHeight: 0.5,
-        closeGapType: 0, // { "NONE" : 0, "ABOVE" : 1, "BELOW" : 2 }
+        closeGapType: params.getNumber("closeGapType", 0), // { "NONE" : 0, "ABOVE" : 1, "BELOW" : 2 }
 
         rebuild: function () {
-          rebuildTerrain();
-        }
+          rebuild();
+        },
+
+        clearPathSegments: params.getNumber("clearPathSegments", true)
       },
       GUP
     );
@@ -114,7 +120,7 @@
     }
 
     function findHeightFaceIntersectionLine(xIndex, yIndex, heightFace, heightValue) {
-      var heightValueA = heightFace[0][0];
+      var heightValueA = heightFace[0][0]; // value at (x,y)
       var heightValueB = heightFace[1][0];
       var heightValueC = heightFace[1][1];
       var heightValueD = heightFace[0][1];
@@ -122,24 +128,24 @@
       var points = [];
       if (isBetween(heightValueA, heightValueB, heightValue)) {
         var lerpValueByHeight = getLerpRatio(heightValueA, heightValueB, heightValue);
-        points.push({ x: lerp(xIndex - 1, xIndex, lerpValueByHeight), y: yIndex - 1 });
+        points.push(new Vertex(lerp(xIndex, xIndex + 1, lerpValueByHeight), yIndex));
       }
       if (isBetween(heightValueB, heightValueC, heightValue)) {
         var lerpValueByHeight = getLerpRatio(heightValueB, heightValueC, heightValue);
-        points.push({ x: xIndex, y: lerp(yIndex - 1, yIndex, lerpValueByHeight) });
+        points.push(new Vertex(xIndex + 1, lerp(yIndex, yIndex + 1, lerpValueByHeight)));
       }
       if (isBetween(heightValueC, heightValueD, heightValue)) {
         var lerpValueByHeight = getLerpRatio(heightValueC, heightValueD, heightValue);
-        points.push({ x: lerp(xIndex, xIndex - 1, lerpValueByHeight), y: yIndex });
+        points.push(new Vertex(lerp(xIndex + 1, xIndex, lerpValueByHeight), yIndex + 1));
       }
       if (isBetween(heightValueD, heightValueA, heightValue)) {
         var lerpValueByHeight = getLerpRatio(heightValueD, heightValueA, heightValue);
-        points.push({ x: xIndex - 1, y: lerp(yIndex, yIndex - 1, lerpValueByHeight) });
+        points.push(new Vertex(xIndex, lerp(yIndex + 1, yIndex, lerpValueByHeight)));
       }
 
       if (points.length >= 2) {
-        var startPoint = convertCoords2Pos(points[0].x, points[0].y);
-        var endPoint = convertCoords2Pos(points[1].x, points[1].y);
+        var startPoint = points[0]; // convertCoords2Pos(points[0].x, points[0].y);
+        var endPoint = points[1]; // convertCoords2Pos(points[1].x, points[1].y);
         // pathSegments.push(new Line(startPoint, endPoint));
         if (points.length > 2) {
           console.log("Ooops, detected more than 2 points on one face at ", xIndex, yIndex);
@@ -152,24 +158,22 @@
       }
     }
 
+    // Array<GenericPath>
+    var pathSegments = [];
+
     /**
      * The re-drawing function.
      */
-    var redraw = function (draw, fill) {
-      // Draw out stuff
-      fill.rect(bounds2D.min, bounds2D.width, bounds2D.height, "rgba(0,0,0,0.75)");
-      draw.rect(bounds2D.min, bounds2D.width, bounds2D.height, "green", 2);
-
-      var pathSegments = [];
+    var rebuildPlotPlane = function (draw, fill) {
+      //   var pathSegments = [];
+      if (config.clearPathSegments) {
+        pathSegments = [];
+      }
 
       // Find a level to "splice" the mesh, here at middle of Min/Max
       var medianHeight =
         terrainGeneration._minHeight + (terrainGeneration._maxHeight - terrainGeneration._minHeight) * config.sliceHeight;
       console.log("medianHeight", medianHeight);
-
-      var LO_COLOR = Color.parse("#0000ff");
-      var HI_COLOR = Color.parse("#ff0000");
-      console.log("LO_COLOR", LO_COLOR, "HI_COLOR", HI_COLOR);
 
       // A face element:
       // (x-1,y-1)      (x,y-1)
@@ -182,34 +186,105 @@
         [0, 0],
         [0, 0]
       ];
-      for (var y = 1; y < terrainGeneration.ySegmentCount; y++) {
-        for (var x = 1; x < terrainGeneration.xSegmentCount; x++) {
-          terrainGeneration.getHeightFace4At(x - 1, y - 1, heightFace);
-          var heightValueA = heightFace[0][0];
-          //   var heightValueB = heightFace[1][0];
-          //   var heightValueC = heightFace[1][1];
-          //   var heightValueD = heightFace[0][1];
-
-          // Draw a point at (x,y) to indicate the height
-          var pointPosition = convertCoords2Pos(x - 1, y - 1);
-          var heightRatio =
-            (heightValueA - terrainGeneration._minHeight) / (terrainGeneration._maxHeight - terrainGeneration._minHeight);
-          var color = LO_COLOR.clone().interpolate(HI_COLOR, heightRatio);
-          fill.point(pointPosition, color.cssRGB(), 2);
-
+      for (var y = 0; y + 1 < terrainGeneration.ySegmentCount; y++) {
+        for (var x = 0; x + 1 < terrainGeneration.xSegmentCount; x++) {
+          terrainGeneration.getHeightFace4At(x, y, heightFace);
           var line = findHeightFaceIntersectionLine(x, y, heightFace, medianHeight);
           if (line) {
-            draw.line(line.a, line.b, "orange", 2);
+            pathSegments.push(new GenericPath(line));
           }
         }
       }
+
+      // Collect value above/below on the y axis
+      var xExtremes = [0, terrainGeneration.xSegmentCount - 1];
+      for (var i = 0; i < xExtremes.length; i++) {
+        var x = xExtremes[i];
+        for (var y = 0; y + 1 < terrainGeneration.ySegmentCount; y++) {
+          var heightValueA = terrainGeneration.getHeightValueAt(x, y);
+          var heightValueB = terrainGeneration.getHeightValueAt(x, y + 1);
+
+          if (config.closeGapType === CLOSE_GAP_TYPE_ABOVE) {
+            if (heightValueA >= medianHeight && heightValueB >= medianHeight) {
+              //  Both above
+              var line = new Line(new Vertex(x, y), new Vertex(x, y + 1));
+              pathSegments.push(new GenericPath(line));
+            } else if (heightValueA >= medianHeight && heightValueB < medianHeight) {
+              // Only one of both (first) is above -> interpolate to find exact intersection point
+              var lerpValueByHeight = getLerpRatio(heightValueA, heightValueB, medianHeight);
+              var interpLine = new Line(new Vertex(x, y), new Vertex(x, lerp(y, y + 1, lerpValueByHeight)));
+              pathSegments.push(new GenericPath(interpLine));
+            } else if (heightValueA < medianHeight && heightValueB >= medianHeight) {
+              // Only one of both (second) is above -> interpolate to find exact intersection point
+              var lerpValueByHeight = getLerpRatio(heightValueA, heightValueB, medianHeight);
+              var interpLine = new Line(new Vertex(x, lerp(y, y + 1, lerpValueByHeight)), new Vertex(x, y + 1));
+              pathSegments.push(new GenericPath(interpLine));
+            }
+          } else if (config.closeGapType === CLOSE_GAP_TYPE_BELOW) {
+            if (heightValueA < medianHeight && heightValueB < medianHeight) {
+              // Both below
+            }
+          }
+        }
+      }
+      var yExtremes = [0, terrainGeneration.ySegmentCount - 1];
     };
 
     var terrainGeneration = new TerrainGeneration("three-canvas", 16, 16);
     var modal = new Modal();
 
+    var rebuild = function () {
+      rebuildTerrain();
+      rebuildPlotPlane();
+    };
+
     var rebuildTerrain = function () {
       terrainGeneration.rebuild(Object.assign({}, config));
+    };
+
+    /**
+     * Build the contour from the current terrain data.
+     */
+    var redraw = function (draw, fill) {
+      // Draw out stuff
+      fill.rect(bounds2D.min, bounds2D.width, bounds2D.height, "rgba(0,0,0,0.75)");
+      draw.rect(bounds2D.min, bounds2D.width, bounds2D.height, "green", 2);
+
+      //   var pathSegments = [];
+
+      // Find a level to "splice" the mesh, here at middle of Min/Max
+      var medianHeight =
+        terrainGeneration._minHeight + (terrainGeneration._maxHeight - terrainGeneration._minHeight) * config.sliceHeight;
+      console.log("medianHeight", medianHeight);
+
+      var LO_COLOR = Color.parse("#0000ff");
+      var HI_COLOR = Color.parse("#ff0000");
+      console.log("LO_COLOR", LO_COLOR, "HI_COLOR", HI_COLOR);
+
+      for (var y = 0; y < terrainGeneration.ySegmentCount; y++) {
+        for (var x = 0; x < terrainGeneration.xSegmentCount; x++) {
+          //   terrainGeneration.getHeightFace4At(x, y, heightFace);
+
+          // Draw a point at (x,y) to indicate the height
+          //   var heightValueA = heightFace[0][0];
+          var heightValueA = terrainGeneration.getHeightValueAt(x, y, false);
+          var pointPosition = convertCoords2Pos(x, y);
+          var heightRatio =
+            (heightValueA - terrainGeneration._minHeight) / (terrainGeneration._maxHeight - terrainGeneration._minHeight);
+          var color = LO_COLOR.clone().interpolate(HI_COLOR, heightRatio);
+          fill.point(pointPosition, color.cssRGB(), 2);
+        }
+      }
+
+      for (var i in pathSegments) {
+        var segment = pathSegments[i];
+        // draw.line(line.a, line.b, "orange", 2);
+        for (var j in segment.segments) {
+          // By construction we now this must be a line (cannot be a curve or so as we didn't add one)
+          var line = segment.segments[j];
+          draw.line(convertCoords2Pos(line.a.x, line.a.y), convertCoords2Pos(line.b.x, line.b.y), "orange", 2);
+        }
+      }
     };
 
     // +---------------------------------------------------------------------------------
@@ -228,18 +303,21 @@
       fold0.add(config, "wireframe").onChange( function() { rebuildTerrain() } ).name('wireframe').title('Display the mesh as a wireframe model.');
 
       // prettier-ignore
-      gui.add(config, "sliceHeight").min(0.0).max(1.0).onChange( function() { pb.redraw(); } ).name('sliceHeight').title('Where to slice the current terrain model.');
+      gui.add(config, "sliceHeight").min(0.0).max(1.0).onChange( function() { rebuildPlotPlane(); pb.redraw(); } ).name('sliceHeight').title('Where to slice the current terrain model.');
       // prettier-ignore
-      gui.add(config, "closeGapType", { "None" : 0, "Above" : 1, "Below" : 2 } ).onChange( function() { pb.redraw(); } ).name('closeGapType').title('Close gap above, below or not at all.');
+      gui.add(config, "closeGapType", { "None" : CLOSE_GAP_TYPE_NONE, "Above" : CLOSE_GAP_TYPE_ABOVE, "Below" : CLOSE_GAP_TYPE_BELOW } ).onChange( function() { rebuildPlotPlane(); pb.redraw(); } ).name('closeGapType').title('Close gap above, below or not at all.');
 
-      var fold1 = gui.addFolder("Export");
       // prettier-ignore
-      fold1.add(config, "exportSTL").name('STL').title('Export an STL file.');
+      gui.add(config, "clearPathSegments").onChange( function() { rebuildPlotPlane(); pb.redraw(); } ).name('clearPathSegments').title('Clear path buffer on each rebuild cycle (default=true).');
+
+      //   var fold1 = gui.addFolder("Export");
+      //   // prettier-ignore
+      //   fold1.add(config, "exportSTL").name('STL').title('Export an STL file.');
     }
 
     pb.config.postDraw = redraw;
     // randomPoints(true, false, false); // clear ; no full cover ; do not redraw
-    rebuildTerrain();
+    rebuild();
     pb.redraw();
   });
 })(window);